fix: coc制卡上限和刷屏频率配置无法正确加载
PR #261 #258
